backend/requirements.txt
pyjwt
uvicorn
fastapi
yagmail
pymysql
pydantic
sqlalchemy
email-validator
python-multipart
python-jose[cryptography]
